requirements.txt
# Moore Machine In-Context Learning Project

## Core Dependencies
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
pandas>=1.5.0
networkx>=3.0
pyyaml>=6.0
tqdm>=4.65.0

# Optional dependencies for enhanced functionality
wandb>=0.15.0  # For experiment tracking
jupyter>=1.0.0  # For notebook analysis
scikit-learn>=1.3.0  # For additional metrics

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.3.0